Replaces maven javadoc tags with annotations
