+++ .gitignore
@@ -1,3 +1,4 @@
+*.checkstyle
 *.swp
 .project
 .settings
