+++ pitest-maven/pom.xml
@@ -50,12 +50,23 @@
 				<version>3.4</version>
 				<executions>
 					<execution>
-						<id>generated-helpmojo</id>
+						<id>default-descriptor</id>
+						<phase>process-classes</phase>
+					</execution>
+					<execution>
+						<id>help-goal</id>
 						<goals>
 							<goal>helpmojo</goal>
 						</goals>
 					</execution>
 				</executions>
+			</plugin>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-javadoc-plugin</artifactId>
+				<configuration>
+					<additionalparam>-Xdoclint:none</additionalparam>
+				</configuration>
 			</plugin>
 		</plugins>
 	</build>
@@ -75,6 +86,12 @@
 			<groupId>org.apache.maven</groupId>
 			<artifactId>maven-plugin-api</artifactId>
 			<version>${maven.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>org.apache.maven.plugin-tools</groupId>
+			<artifactId>maven-plugin-annotations</artifactId>
+			<version>3.4</version>
+			<scope>provided</scope><!-- annotations are needed only to build the plugin -->
 		</dependency>
 		<dependency>
 			<groupId>org.apache.maven.reporting</groupId>
