+++ pitest-maven/src/main/java/org/pitest/maven/report/PitReportMojo.java
@@ -19,6 +19,10 @@ import java.util.List;
 import java.util.Locale;
 
 import org.apache.maven.doxia.siterenderer.Renderer;
+import org.apache.maven.plugins.annotations.Component;
+import org.apache.maven.plugins.annotations.LifecyclePhase;
+import org.apache.maven.plugins.annotations.Mojo;
+import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.project.MavenProject;
 import org.apache.maven.reporting.AbstractMavenReport;
 import org.apache.maven.reporting.MavenReportException;
@@ -28,32 +32,21 @@ import org.pitest.util.PitError;
 
 /**
  * Generates a report of the pit mutation testing.
- * 
- * @goal report
- * @phase site
  */
+@Mojo(name = "report", defaultPhase = LifecyclePhase.SITE)
 public class PitReportMojo extends AbstractMavenReport {
 
-  /**
-   * @component
-   * @required
-   * @readonly
-   */
+  @Component
   private Renderer                siteRenderer;
 
-  /**
-   * @parameter default-value="${project}"
-   * @required
-   * @readonly
-   */
+  @Parameter(defaultValue = "${project}", required = true, readonly = true)
   private MavenProject            project;
 
   /**
    * When set to true, indicates that generation of the site report should be
    * skipped.
-   * 
-   * @parameter default-value="false" expression="${pit.report.skip}"
    */
+  @Parameter(property = "pit.report.skip", defaultValue = "false")
   private boolean                 skip;
 
   /**
@@ -61,10 +54,8 @@ public class PitReportMojo extends AbstractMavenReport {
    * goal. If timestampedReports is true (the default), then the actual reports
    * will be contained in a subdirectory within this directory. If
    * timestampedReports is false, the actual reports will be in this directory.
-   * 
-   * @parameter default-value="${project.build.directory}/pit-reports"
-   *            expression="${reportsDirectory}"
    */
+  @Parameter(property = "reportsDirectory", defaultValue = "${project.build.directory}/pit-reports")
   private File                    reportsDirectory;
 
   /**
@@ -73,27 +64,22 @@ public class PitReportMojo extends AbstractMavenReport {
    * HTML report generated by the mutationCoverage goal into the site directory.
    * However, the plan is to accept any source format that can be specified in
    * the outputFormats parameter of the mutationCoverage goal.
-   * 
-   * @parameter default-value="HTML"
-   *            expression="${pit.report.sourceDataFormats}"
    */
+  @Parameter(property = "pit.report.sourceDataFormats", defaultValue = "HTML")
   private List<String>            sourceDataFormats;
 
   /**
    * Determines the name of the report that is displayed under the site's
    * Project Reports section
-   * 
-   * @parameter default-value="PIT Test Report" expression="${pit.report.name}"
    */
+  @Parameter(property = "pit.report.name", defaultValue = "PIT Test Report")
   private String                  siteReportName;
 
   /**
    * Determines the description of the report that is displayed under the
    * description of the site's Project Reports section
-   * 
-   * @parameter default-value="Report of the pit test coverage"
-   *            expression="${pit.report.description}"
    */
+  @Parameter(property = "pit.report.description", defaultValue = "Report of the pit test coverage")
   private String                  siteReportDescription;
 
   /**
@@ -103,8 +89,8 @@ public class PitReportMojo extends AbstractMavenReport {
    * parameter is set to "foo", then the pit reports will be located in
    * ${project.reporting.outputDirectory}/foo.
    * 
-   * @parameter default-value="pit-reports" expression="${pit.report.outputdir}"
    */
+  @Parameter(property = "pit.report.outputdir", defaultValue = "pit-reports")
   private String                  siteReportDirectory;
 
   private ReportGenerationManager reportGenerationManager;
